requirements.txt
torch
torchvision
transformers
sentence-transformers
faiss-cpu
pandas
numpy
pillow
opencv-python
scikit-learn
nltk
sacrebleu
matplotlib
seaborn
gradio
streamlit
langchain
tqdm
sentencepiece
kaggle
kagglehub
deep-translator

